% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_jitter.R
\name{add_jitter}
\alias{add_jitter}
\title{Add jitter to any points that are overlapping at same xy-coordinates}
\usage{
add_jitter(df, col1 = "x", col2 = "y", jitter_fun = arrange_points_2d, ...)
}
\arguments{
\item{df}{data frame with points to hard-add jitter to}

\item{col1}{column name of x-coord}

\item{col2}{column name of y-coord}

\item{jitter_fun}{arrange_points_2d or arrange_points_2d_circle
or any function that returns a matrix or data frame with columns
1 and 2 containing cartesian coordinates}

\item{...}{arguments to jitter_fun}
}
\value{
df with jittered coordinates
}
\description{
Add jitter to any points that are overlapping at same xy-coordinates
}
\examples{
df <- data.frame(x = round(rnorm(10)), y = round(rnorm(10)))
n <- 30  # Number of times to repeat
df2 <- do.call(rbind, replicate(n, df, simplify = FALSE))
ggplot2::ggplot(df2, ggplot2::aes(x,y)) + ggplot2::geom_point()
df3 <- add_jitter(df2)
ggplot2::ggplot(df3, ggplot2::aes(x,y)) + ggplot2::geom_point()
df4 <- add_jitter(df2, jitter_fun = arrange_points_2d_circle)
ggplot2::ggplot(df4, ggplot2::aes(x,y)) + ggplot2::geom_point()
}
