% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_points_2d.R
\name{arrange_points_2d}
\alias{arrange_points_2d}
\title{Evenly arrange n points in 2D space in square shape}
\usage{
arrange_points_2d(
  n,
  height = 1,
  width = 1,
  scale = FALSE,
  center = TRUE,
  center_incomplete_rows = TRUE
)
}
\arguments{
\item{n}{number of points}

\item{height}{spread in y-direction}

\item{width}{spread in x-direction}

\item{scale}{do scale position of arranged points?}

\item{center}{center their position?}

\item{center_incomplete_rows}{center rows which are not filled completely}
}
\value{
matrix with 2 columns
}
\description{
May be used to add an ordered jitter to overlapping points.
}
\examples{
plot(arrange_points_2d(100))
plot(arrange_points_2d(100, height = 10, width = 5, center = FALSE))
plot(arrange_points_2d(195, height = 10, width = 5, center_incomplete_rows = TRUE))
}
