% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling_floor2.R
\name{ceiling2}
\alias{ceiling2}
\title{Ceiling to a given decimal place}
\usage{
ceiling2(x, accuracy = 1, fun = base::ceiling)
}
\arguments{
\item{x}{numeric vector}

\item{accuracy}{decimal accuracy; as decimal like 0.1 or 0.01 or as digits
like 1, 2 as in base::round()}

\item{fun}{function}
}
\value{
numeric vector
}
\description{
Ceiling to a given decimal place
}
\examples{
ceiling2(rnorm(20))
}
