% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clr.R
\name{clr}
\alias{clr}
\title{Title}
\usage{
clr(x, margin = 1)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{margin}{if matrix: iterate over rows or cols?}
}
\value{
vector or matrix
}
\description{
compositions::clr(c(0.2,0.5,0.3,0))
}
\examples{
x <- c(NA,0.2,NA,0.3,0,0.5,0)
clr(x)
compositions::clr(x)
}
