% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_corrcoeff.R
\name{combine_corrcoeff}
\alias{combine_corrcoeff}
\title{Combine / average correlation coefficients}
\usage{
combine_corrcoeff(x)
}
\arguments{
\item{x}{vector of correlation coefficients}
}
\value{
average coeff, numeric
}
\description{
Calculates Fisher z average: tanh(mean(atanh(x)))
}
