% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_pvalues.R
\name{combine_pvalues}
\alias{combine_pvalues}
\title{Average p-values with fisher of stouffer method}
\usage{
combine_pvalues(pvalues, method = c("fisher", "stouffer"), sample_sizes = NULL)
}
\arguments{
\item{pvalues}{vector of p-values}

\item{method}{method}

\item{sample_sizes}{for stouffer method:
optional samples size to add weights to p-vals}
}
\value{
numeric: averaged p-value
}
\description{
Check packages metap or poolr
}
