% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_matrices.R
\name{compare_matrices}
\alias{compare_matrices}
\title{Find matching rows/cols of matrices by hashing}
\usage{
compare_matrices(
  x1,
  x2,
  margin = 2,
  match_names = F,
  chunk_size = 500,
  ncores = 0
)
}
\arguments{
\item{x1}{matrix1}

\item{x2}{matrix2}

\item{margin}{1 = rows, 2 = columns}

\item{match_names}{filter and order by intersection of x1, x2 (in opposite of
margin)}

\item{chunk_size}{split wise calculation to reduce memory footprint (with
sparse matrices)}
}
\value{
data frame
}
\description{
Find matching rows/cols of matrices by hashing
}
\examples{
x1 <- matrix(as.integer(rnorm(100)), ncol = 20)
x2 <- matrix(as.integer(rnorm(100)), ncol = 20)
df <- compare_matrices(x1,x2)
A <- matrix(rnorm(15), nrow = 5, ncol = 3)
rownames(A) <- paste0("row", 1:5)
colnames(A) <- paste0("A_col", 1:3)

# Second matrix: 5 rows, 4 columns
B <- matrix(rnorm(20), nrow = 5, ncol = 4)
rownames(B) <- paste0("row", c(3:7))  # overlaps partially: row3, row4, row5
colnames(B) <- paste0("B_col", 1:4)
compare_matrices(A,B, match_names = T)
}
