% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contour_est.R
\name{contour_est}
\alias{contour_est}
\title{2D contour estimation}
\usage{
contour_est(x, y, n = 100, levels = NULL, ...)
}
\arguments{
\item{x}{matrix or x values, when matrix: col1 becomes x, col2 becomes y}

\item{y}{optional: y values when x is not a matrix}

\item{n}{number of bins, passed to MASS::kde2d}

\item{levels}{passed to grDevices::contourLines, if NULL determined by
algorithm}

\item{...}{arguments to MASS::kde2d}
}
\value{
data frame
}
\description{
2D contour estimation
}
\examples{
contour_est(data.frame(x = rnorm(1000), y = rnorm(1000)))
}
