% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_subset_word_groups.R
\name{count_subset_word_groups}
\alias{count_subset_word_groups}
\title{Count common word groups from a vector of characters}
\usage{
count_subset_word_groups(
  x,
  split = " ",
  collapse = split,
  min_len = 1,
  consecutive_only = T
)
}
\arguments{
\item{x}{character vector}

\item{split}{where to split x}

\item{collapse}{how to collapse subset groups}
}
\value{
data frame
}
\description{
Count common word groups from a vector of characters
}
\examples{
x <- c("this is a vector", "this is a list", "this is a vector group", "a list is what")
df1 <- count_subset_word_groups(x)
df2 <- count_subset_word_groups(x, collapse = "_")
}
