% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling_floor2.R
\name{decimals_adaptive}
\alias{decimals_adaptive}
\title{Adapt decimal places to input values}
\usage{
decimals_adaptive(x)
}
\arguments{
\item{x}{numeric vector}
}
\value{
integer of suggested meaningful decimal places to round to
}
\description{
Largest absolute value above 100: 0 digits.
Above 1: 1 digit.
Below 1: (largest decimal place with differences) + 1
}
\examples{
decimals_adaptive(c(109,15))
decimals_adaptive(x = c(10.5,15))
decimals_adaptive(x = c(10,15))
decimals_adaptive(x = c(1.5,5))
decimals_adaptive(x = c(1,5))
decimals_adaptive(x = c(0.4,0.2))
y <- decimals_adaptive(x = c(0.456,0.265))
# use for rounding
z <- round(x = c(0.456,0.265), digits = y)
}
