% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density_est.R
\name{density_est}
\alias{density_est}
\title{2D or 3D density estimation}
\usage{
density_est(x, y, z, n = 100, type = c("2D", "3D"), ...)
}
\arguments{
\item{x}{matrix or x values, when matrix: col1 becomes x, col2 becomes y}

\item{y}{optional: y values when x is not a matrix}

\item{z}{optional: z values when x is not a matrix}

\item{n}{number of bins, passed to MASS::kde2d}

\item{type}{2D or 3D estimation}

\item{...}{arguments to MASS::kde2d}
}
\value{
list with raw and 0-to-1-scaled density values
}
\description{
2D or 3D density estimation
}
\examples{
density_est(data.frame(x = rnorm(1000), y = rnorm(1000), z = rnorm(1000)))
}
