% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdf_eval.R
\name{ecdf_eval}
\alias{ecdf_eval}
\title{Evaluate (groupwise) ECDF at a given cutoff (reverse quantile)}
\usage{
ecdf_eval(df, valuecol, cutoff, groupcol = NULL, strict = FALSE, na_rm = TRUE)
}
\arguments{
\item{df}{data frame.}

\item{valuecol}{numeric column to build the ECDF(s) on}

\item{cutoff}{numeric cutoff at which to evaluate ECDF}

\item{groupcol}{optional grouping column}

\item{strict}{<cutoff or <=cutoff}

\item{na_rm}{remove NA?}
}
\value{
A tibble with one row per group and cutoff
}
\description{
Related to quantile_filter.
}
