% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_points_2d.R
\name{ellipse_points_2d}
\alias{ellipse_points_2d}
\title{Sample points on ellipse with cov matrix}
\usage{
ellipse_points_2d(
  mu = c(0, 0),
  covmat = matrix(c(1, 0, 0, 1), nrow = 2),
  n = 100,
  c = 1,
  tol = 1e-10
)
}
\arguments{
\item{mu}{center coordinates}

\item{covmat}{covariance matrix}

\item{n}{number of points}

\item{c}{radius or size factor}
}
\value{
data frame
}
\description{
Sample points on ellipse with cov matrix
}
\examples{
library(ggplot2)
ggplot(mapping = aes(x,y)) +
geom_point(data = ellipse_points_2d()) +
    geom_point(data = ellipse_points_2d(c = 3), color = "firebrick") +
    geom_point(data = ellipse_points_2d(mu = c(1,1), covmat = c(4,0,0,1)), color = "cornflowerblue") +
    geom_point(data = ellipse_points_2d(mu = c(-1,-1), covmat = c(4,3,0,1)), color = "forestgreen") +
    geom_point(data = ellipse_points_2d(mu = c(-1,-1), covmat = c(4,0,3,1)), color = "hotpink") +
    geom_point(data = ellipse_points_2d(mu = c(1,-1), covmat = c(4,2,2,2)), color = "grey60") +
    coord_equal()
}
