% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_num_order.R
\name{factor_num_order}
\alias{factor_num_order}
\title{Order character by contained numerics}
\usage{
factor_num_order(x, return = c("levels", "factor"))
}
\arguments{
\item{x}{character vector numbers in it}

\item{return}{type to return. levels: ordered levels of x. factor: x as factor
with ordered levels}
}
\value{
character or factor
}
\description{
Order character by contained numerics
}
\examples{
# one num per character
factor_num_order(c("C3", "C1", "C2"))
# multiple nums
factor_num_order(c("C9C3", "C9C1", "C9C2"))
factor_num_order(c("C9x8C3", "C9x8C1", "C9x8C2"))
# diff numbers of nums across x
factor_num_order(c("C9x8C3", "C9x8C1", "C9x8C2v7"))
}
