% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_duplicate_rows.R
\name{filter_duplicate_rows}
\alias{filter_duplicate_rows}
\title{Get all duplicate rows for inspection}
\usage{
filter_duplicate_rows(x, cols = names(x))
}
\arguments{
\item{x}{data frame}

\item{cols}{column names to check for duplicates}
}
\value{
data frame
}
\description{
Uses dplyr::add_count. Better than any solution with duplicated()
}
\examples{
df <- data.frame(x = c(letters, letters), y = c(LETTERS, LETTERS))
filter_duplicate_rows(df)
}
