% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_edge_run.R
\name{find_edge_run_indices}
\alias{find_edge_run_indices}
\title{Find stretches of leading or trailing sequences below a threshold}
\usage{
find_edge_run_indices(x, threshold, fun = "<")
}
\arguments{
\item{x}{numeric vector}

\item{threshold}{upper threshold}

\item{fun}{comparison between x and threshold, x comes first: x < threshold}
}
\value{
named numeric vector
}
\description{
Find stretches of leading or trailing sequences below a threshold
}
\examples{
x <- c(0.5, 1.2, 2.1, 3.5, 0.8, 0.3, 4.2, 5.1, 0.1, 0.2)
find_edge_run_indices(x, threshold = 1)
x <- c(NA,NA, 0.5, 1.2, 2.1, 3.5, 0.8, 0.3, 4.2, 5.1,1, 0.1, 0.2,NA)
find_edge_run_indices(x, threshold = 6, fun = "<")
#' x <- c(NA,NA, 0.5, 1.2, 2.1, 3.5, 0.8, 0.3, 4.2, 5.1,1, 0.1, 0.2,NA)
find_edge_run_indices(x, threshold = 5, fun = "<")
x <- c(NA,NA, 0.5, 1.2, 2.1, 3.5, NA, 0.8, 0.3, 4.2, 5.1,1, 0.1, 0.2,NA)
find_edge_run_indices(x, threshold = 5, fun = "<")
}
