% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_char_sep.R
\name{find_sep}
\alias{find_sep}
\title{Find a valid separator which is not present in any provided character}
\usage{
find_sep(x, sep = "_")
}
\arguments{
\item{x}{character vector}

\item{sep}{character}
}
\value{
character
}
\description{
Sep is repeated until it is unique (not found in any of x).
}
\examples{
x <- c("my_name_is", "coward__harpendale")
find_sep(x = x)
}
