% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_local_extrema_inds.R
\name{findlocmax}
\alias{findlocmax}
\title{Title}
\usage{
findlocmax(
  x,
  min_gap = 0,
  threshold = -Inf,
  type = c("max", "min"),
  min_mono_width = 1,
  sizes = c(1, 2, 3, 4, 5),
  strictly_mono = F
)
}
\arguments{
\item{type}{}
}
\description{
See package gcplyr.
}
