% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_contour_2d.R
\name{get_contour_2d}
\alias{get_contour_2d}
\title{Calculate contour lines from x-y coordinates}
\usage{
get_contour_2d(
  x,
  y,
  group = NULL,
  n = NULL,
  scale = T,
  nlevels = 10,
  levels = NULL
)
}
\arguments{
\item{x}{x coordinates or data frame}

\item{y}{y coordinates or empty when x is data frame}

\item{group}{group vector of column name in x}

\item{n}{passed to MASS::kde2d; large: rounder contours}

\item{scale}{scale levels to range 0-1 (for each group separately when
!is.null(group))}

\item{nlevels}{how many contour line levels}

\item{levels}{define contour levels explicitly; easy to use when scale = T
because then choose value between 0-1; if scale = F levels are not predictable
before}
}
\value{
data frame to use with ggplot2::geom_path
}
\description{
Calculate contour lines from x-y coordinates
}
\examples{
df1 <- data.frame(x = rnorm(2000, mean = -10, sd = 3),
                  y = rnorm(2000, mean = -10, sd = 1),
                  group = "1")
df2 <- data.frame(x = rnorm(2000, mean = 10, sd = 1),
                  y = rnorm(2000, mean = 10, sd = 2),
                  group = "2")
df <- rbind(df1, df2)
cdf <- get_contour_2d(x = df, group = "group", levels = 0.2)
ggplot(df, aes(x, y)) +
    geom_point(alpha = 0.3, aes(color = group)) +
    geom_path(data = cdf, aes(x, y, group = combined_group))
}
