% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repel_coords.R
\name{get_repel_coords}
\alias{get_repel_coords}
\title{Get coordinates of repelled labels from ggrepel}
\usage{
get_repel_coords(obj, width = 1, height = 1, max.overlaps = Inf, ...)
}
\arguments{
\item{obj}{data frame with first columns being xy-coords and third column
labels}

\item{width}{passed to grid::viewport, may be subject to iteration to obtain
optimal repelled coordinates}

\item{height}{passed to grid::viewport}

\item{max.overlaps}{ggrepel::geom_text_repel argument}

\item{...}{arguments to ggrepel::geom_text_repel}
}
\value{
data frame
}
\description{
Adapted from https://github.com/slowkow/ggrepel/issues/24. Labels are
ploted very late in the plotting process, i.e. their positions are not
pre-computed or so. Hence, it is not straight forward to extract their
repelled positions.
}
\examples{
library(ggplot2)
df <- data.frame(x=rnorm(10), y=rnorm(10), z = letters[1:10])
gg <-
    ggplot(df, aes(x = x, y = y, label = z)) +
    geom_point() +
    ggrepel::geom_label_repel() +
    geom_text(data = get_repel_coords(obj = df), color = "hotpink")
gg
}
