% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_get_axis_text.R
\name{gg_get_axis_text}
\alias{gg_get_axis_text}
\title{Get axis breaks in order from ggplot}
\usage{
gg_get_axis_text(ggobj)
}
\arguments{
\item{ggobj}{ggplot object}
}
\value{
list of x and y breaks
}
\description{
Get axis breaks in order from ggplot
}
