% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_get_theme_ele.R
\name{gg_get_theme_element}
\alias{gg_get_theme_element}
\title{Get theme element from ggplot2 object}
\usage{
gg_get_theme_element(ggobj, element = "plot.background")
}
\arguments{
\item{ggobj}{ggplot2 object}

\item{element}{theme element name}
}
\description{
Get theme element from ggplot2 object
}
