% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_inject_theme_element.R
\name{gg_inject_theme_element}
\alias{gg_inject_theme_element}
\title{Alter a specific ggplot2 theme element}
\usage{
gg_inject_theme_element(theme_args, elem, elem_sub, value)
}
\arguments{
\item{theme_args}{list of ggplot theme elements}

\item{elem}{element name}

\item{elem_sub}{element subtype}

\item{value}{value to write}
}
\value{
theme_args list with new value
}
\description{
Overwriting other parts of elem except elem_sub is avoided.
}
\examples{
theme_args <- ggplot2::theme_get()
theme_args <- gg_inject_theme_element(theme_args, "axis.text.x", "face", value = "italic")
}
