% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_points_2d.R
\name{givens_rotation}
\alias{givens_rotation}
\title{Get rotation matrix for nD}
\usage{
givens_rotation(n, i, j, theta)
}
\arguments{
\item{n}{total dimensions}

\item{i}{dim1 to rotate}

\item{j}{dim2 to rotate}

\item{theta}{angle in radians}
}
\description{
Rotates the plane of 2 dimensions
}
\examples{
# Combine multiple rotations
R <- diag(4)
R <- givens_rotation(4, 1, 2, pi/6) \%*\% R
R <- givens_rotation(4, 2, 3, pi/4) \%*\% R
R <- givens_rotation(4, 3, 4, pi/3) \%*\% R
}
