% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_tight.R
\name{gt_tight}
\alias{gt_tight}
\title{Make a tight gt table by default}
\usage{
gt_tight(df, font_size = 11, font_family = "Arial")
}
\arguments{
\item{df}{data frame}

\item{font_size}{font size}

\item{font_family}{family}
}
\value{
gt table
}
\description{
Make a tight gt table by default
}
