% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image_convert.R
\name{image_convert}
\alias{image_convert}
\title{Convert image between formats}
\usage{
image_convert(path, format = "jpeg", path_save = NULL, ...)
}
\arguments{
\item{path}{file path on disk}

\item{format}{target format, see magick::image_write}

\item{path_save}{save path of converted image, if NuLL saved to same
folder as the original}

\item{...}{arguments to magick::image_write}
}
\value{
nothing
}
\description{
Converted image is saved in same location as the original (default),
under same name but with new file extension.
}
\examples{
\dontrun{
image_convert(path/to/image.heic, format = "png")

}
}
