% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_to_pptx.R
\name{img_to_pptx}
\alias{img_to_pptx}
\title{Images to pptx file}
\usage{
img_to_pptx(img_paths, pptx_path)
}
\arguments{
\item{img_paths}{paths to files on disk, filtered for pdf, png, tiff, bmp,
jpg, jpeg}

\item{pptx_path}{path to save pptx to}
}
\value{
nothing
}
\description{
Images to pptx file
}
\examples{
\dontrun{
img_to_pptx(imgs[-1], pptx_path = "/Users/chris/Documents/2025_DRFZ_Daten/20250823_GZMB_GZMK_induction/R_images/img_to_pptx.pptx")
}
}
