% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersect_pairwise.R
\name{intersect_pairwise}
\alias{intersect_pairwise}
\title{Highly efficient pairwise intersect}
\usage{
intersect_pairwise(x, as = c("sparse", "dense"))
}
\arguments{
\item{x}{list of character vectors}

\item{as}{how to return}
}
\value{
matrix
}
\description{
Highly efficient pairwise intersect
}
