% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_pairwise.R
\name{jaccard_pairwise}
\alias{jaccard_pairwise}
\title{Highly efficient pairwise jaccard index}
\usage{
jaccard_pairwise(x, as = c("sparse", "dense"))
}
\arguments{
\item{x}{list of character vectors}

\item{as}{how to return}
}
\value{
matrix
}
\description{
Highly efficient pairwise jaccard index
}
