% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_chunkwise.R
\name{join_chunkwise}
\alias{join_chunkwise}
\title{Join large number of data frames chunkwise}
\usage{
join_chunkwise(
  df_list,
  join_fun = dplyr::full_join,
  join_by,
  chunks = length(df_list)/10,
  progress = F,
  rm_invalid = F
)
}
\arguments{
\item{df_list}{list of data frames}

\item{join_fun}{function to use for joining; must be one of dplyrs functions}

\item{join_by}{column(s) to join by; passed to "by" arg of join_fun}

\item{chunks}{number of chunks}

\item{progress}{print progress bar?}

\item{rm_invalid}{remove invalid list elements (e.g, no data frame or
no join_by column)}
}
\value{
a single data frame
}
\description{
It is quicker with a reasonable number of chunk.
}
