% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kendall_zi_cross.R
\name{kendall_zi_cross}
\alias{kendall_zi_cross}
\title{Kendall's tau for zero-inflated count data between columns of two matrices}
\usage{
kendall_zi_cross(mat1, mat2, mc.cores = 1)
}
\arguments{
\item{mat1}{matrix 1}

\item{mat2}{matrix 2}

\item{mc.cores}{parallel computing?}
}
\value{
matrix of correlation indices
}
\description{
Extension of dismay::kendall_zi.
}
