% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_removal.R
\name{lead_inds}
\alias{lead_inds}
\title{Get indices of leading elements of one kind}
\usage{
lead_inds(x, what = 0)
}
\arguments{
\item{x}{vector of whatever}

\item{what}{element to count}
}
\value{
vector of indices starting at 1
}
\description{
Trivial but included as analogy to trail_inds. Originally intended to
count zeros. So, what = 0 by default.
}
\examples{
lead_inds(c(0,0,1,2,3,4,5))
lead_inds(c("r","r","r","s","t","u","d","i","o"), "r")
lead_inds(rev(c(1,2,3,4,0,0,0,0)))
lead_inds(strsplit("111110001110011001","")[[1]],"1")
lead_inds(c(NA, NA, NA, 0,0,1,1,2,"a"), NA_character_)
}
