% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_depth.R
\name{list_depth}
\alias{list_depth}
\title{Get levels of nesting}
\usage{
list_depth(x)
}
\arguments{
\item{x}{a nested list}
}
\value{
numeric
}
\description{
Get levels of nesting
}
\examples{
list_depth(list(z = list(y = list(x = 3))))
}
