% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_invert.R
\name{list_invert}
\alias{list_invert}
\title{Reorganize a nested list by switching inner and outer layer}
\usage{
list_invert(list, non_inner_list = c("keep", "drop"))
}
\arguments{
\item{list}{a nested list}
}
\value{
list with outer and first inner layer swapped
}
\description{
Reorganize a nested list by switching inner and outer layer
}
\examples{
exam <- list(
  a = list(
    df  = 1,
    vec = 2,
    str = 3
  ),
  b = list(
    df  = 4,
    vec = 5,
    str = 6
  ),
  c = list(
    df  = 7,
    vec = 8,
    str = 9
  ),
  d = list(
    df  = "x",
    vec = "y",
    str = "z",
    num = "w"
  ),
  e = "no_list"
)
inv_exam1 <- list_invert(exam, non_inner_list = "keep")
inv_exam2 <- list_invert(exam, non_inner_list = "drop")
}
