% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_to_binary_matrix.R
\name{list_to_binary_matrix}
\alias{list_to_binary_matrix}
\title{Vector list to membership matrix or incidence matrix}
\usage{
list_to_binary_matrix(x)
}
\arguments{
\item{x}{(named) list of vectors}
}
\value{
sparse binary matrix
}
\description{
Create a membership matrix where 1s indicate membership.
Names of x become rows, unique elements in x become columns.
}
\examples{
# molecular signature genes
msigdb <- scexpr::gsea_get_msigdb()
mat <- list_to_binary_matrix(msigdb$sets)

# prime factors
y <- RcppAlgos::primeFactorizeSieve(1e2)
mat2 <- list_to_binary_matrix(y) # not only 1s
}
