% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_transpose.R
\name{list_transpose}
\alias{list_transpose}
\title{Transpose a List of Equal-Length Vectors}
\usage{
list_transpose(x)
}
\arguments{
\item{x}{list of equal-length vectors}
}
\value{
A list of the same length as the individual input vectors. Each element of the
returned list is a vector containing the corresponding elements across all
input vectors.
}
\description{
Transposes a list of equal-length vectors, effectively swapping rows and columns.
This is analogous to matrix transposition, but applied to a list structure.
}
\details{
The input must be a list where each element is an atomic vector of the same length.
The function returns a new list where the \emph{i}-th element contains the \emph{i}-th
element from each input vector. In other words, it converts a list of row vectors
into a list of column vectors (or vice versa).
}
\examples{
x <- list(
  c(1, 2, 3),
  c(4, 5, 6),
  c(7, 8, 9)
)
list_transpose(x)
}
