% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_unflatten.R
\name{list_unflatten}
\alias{list_unflatten}
\title{Reverse purrr::list_flatten}
\usage{
list_unflatten(x, sep = "_")
}
\arguments{
\item{x}{a list of depth one}

\item{sep}{separator in names of x used in name_spec of purrr::list_flatten}
}
\value{
list of lists
}
\description{
Reverse purrr::list_flatten
}
\examples{
orig <- list(
a = list(
    x = 1,
    y = 2
),
b = list(
    z = 3
)
)
flat <- purrr::list_flatten(orig)
orig2 <- list_unflatten(flat)
identical(orig, orig2)
}
