% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filepath_unique.R
\name{make_filepath_unique}
\alias{make_filepath_unique}
\title{Compare a file path against disk and make it unique by trailing number}
\usage{
make_filepath_unique(path, sep = "_")
}
\arguments{
\item{path}{path/on/disk/text.txt}

\item{sep}{separator before trailing integer}
}
\value{
new path
}
\description{
The resulting path can be used to generate a file on disk without overwriting
an existing file. The trailing integer is incremented from 1 to Inf
}
\examples{
make_filepath_unique(list.files(getwd(), full.names = T)[1])
}
