% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_portable_files.R
\name{make_portable_filename}
\alias{make_portable_filename}
\title{Make file names portable}
\usage{
make_portable_filename(
  x,
  allow = c("-", "."),
  repl = "_",
  urldecode = T,
  make_unique = F,
  repl_empty = "x",
  try_deunicode = F
)
}
\arguments{
\item{x}{character vector of file names}

\item{allow}{values to allow besides alnum}

\item{repl}{replacement for all not permitted values}

\item{urldecode}{also decode URL encoding - the will replace %num
with respective chars}

\item{make_unique}{make filnames unique?}

\item{repl_empty}{replacement for elements that become empty strings during
the procedure}

\item{try_deunicode}{try the deunicode crate from Rust which handles
more characters}
}
\value{
character vector of modified names
}
\description{
That means to make them compatible with many systems by replacing
special non-numeric and non-character values.
}
\examples{
make_portable_filename(x = c(
  "n\u00e4me1",
  "n\u00e6me2",
  "name~3!",
  "name!!4",
  "...name\u00a75___",
  "name/\\\\6",
  "name_\u03b2\u03b37",
  "name{(8)}",
  "name_9\%",
  "",
  "",
  NA))
make_portable_filename(x = c("name\%20_\%9"))
make_portable_filename(x = c("name\%20_\%9"), urldecode = F)
# not everything works
make_portable_filename(x = c("\u5317\u4eb0", "\u3052\u3093\u307e\u3044\u8336",
  "\U0001f984\u2623", "\u2026", "\u1515\u14c7\u14c7"))
# but with deunicode it does
make_portable_filename(x = c("\u5317\u4eb0", "\u3052\u3093\u307e\u3044\u8336",
  "\U0001f984\u2623", "\u2026", "\u1515\u14c7\u14c7"), try_deunicode = T)
}
