% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership_matrix.R
\name{membership_matrix}
\alias{membership_matrix}
\title{Get membership matrix from list of vectors}
\usage{
membership_matrix(x, as = c("sparse", "dense"), verbose = T)
}
\arguments{
\item{x}{list of character vectors}

\item{as}{how to return}

\item{verbose}{print messages?}
}
\value{
(sparse) matrix
}
\description{
Get membership matrix from list of vectors
}
