% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm.R
\name{na_fix_ind}
\alias{na_fix_ind}
\title{Alter a vector of indices according to inserted NA}
\usage{
na_fix_ind(x, x_na, x_inds)
}
\arguments{
\item{x}{vector before addition of NA}

\item{x_na}{vector after addition of NA}

\item{x_inds}{vector of original indices pointing to desired values in x}
}
\value{
adjusted indices vector
}
\description{
The resulting vector of indices points to the same values as before
insertion of NA.
}
\examples{
# initial vector may contain NA
x <- c(NA,1:10)
x_inds <- seq(1,10,2)
na_inds <- c(2,4,6)
# orignal values
x[x_inds]
# insert additional NA
x2 <- na_insert(x, na_inds)
x_inds2 <- na_fix_ind(x, x2, x_inds)
x2[x_inds2] # same as x[x_inds]
}
