% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm.R
\name{na_insert}
\alias{na_insert}
\title{Insert NA at specific indices}
\usage{
na_insert(x, na_inds = NULL)
}
\arguments{
\item{x}{a vector or a list as returned from brathering::na_rm}

\item{na_inds}{indices were NA should be inserted in final vector}
}
\value{
vector
}
\description{
Works on its own but especially in concert with brathering::na_rm.
}
\examples{
x_wo_na <- na_rm(x = c(NA,1,2,NA,3,4,5,NA))
x_wo_na
x_restored <- na_insert(x_wo_na)
x_restored
# na_inds can be tricky as length(x) increases
na_insert(x = c(1:5), na_inds = c(1,3,5,9))
}
