% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_rm.R
\name{na_rm}
\alias{na_rm}
\title{Remove NA from vector and record their indices}
\usage{
na_rm(x, replace = NULL)
}
\arguments{
\item{x}{vector with NA}

\item{replace}{NULL to remove NA, or any value to replace them with}
}
\value{
vector
}
\description{
This may allow to work with the 'cleaned' vector x and subsequently
re-insert the NA.
}
\examples{
x_wo_na <- na_rm(x = c(NA,1,2,NA,3,4,5,NA))
x_wo_na
x_restored <- na_insert(x_wo_na)
x_restored
}
