% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier}
\alias{outlier}
\title{Outlier detection in high dimensional space}
\usage{
outlier(
  df,
  pca = T,
  pcs = 2,
  return = c("outlier", "numeric"),
  methods = c("HDoutliers", "dbscan", "Seurat", "mahalanobis", "Rlof", "mvoutlier",
    "robustbase", "mclust")
)
}
\arguments{
\item{df}{data frame or matrix with numeric columns only}

\item{pca}{run pca on df?}

\item{pcs}{number of components for pca}

\item{return}{instead of outlier classification, return numeric result from
algorithm where possible (Seurat, mahalanobis, Rlof, mclust); for mclust
classification}

\item{methods}{methods for outlier detection}
}
\description{
For Seurat leverage score is calculated. For mclust cluster identities are
calculated followed by heuristic for outlier decision. Leverage score is
fast but may fail with very few observations.
}
\examples{
library(ggplot2)
df <- brathering:::generate_datasets(row_powers = 4, col_seq = 10)[[1]]
um <- fcexpr::ff_calc_umap_tsne(exprs = as.matrix(df))
pc <- fcexpr::ff_calc_pca(exprs = as.matrix(df))[["x"]][,c(1,2)]
outmat <- brathering::outlier(df)

dfres <- dplyr::bind_cols(as.data.frame(um), as.data.frame(pc), as.data.frame(outmat))
dfres <- tidyr::pivot_longer(dfres, -c(UMAP_1, UMAP_2, PC1, PC2), names_to = "method", values_to = "outlier")

ggplot(dfres, aes(UMAP_1, UMAP_2)) +
    geom_point(aes(color = as.factor(outlier))) +
    facet_wrap(vars(method))

ggplot(dfres, aes(PC1, PC2)) +
    geom_point(aes(color = as.factor(outlier))) +
    facet_wrap(vars(method))
}
