% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{outlier1d}
\alias{outlier1d}
\title{Simple methods for outlier detection in 1D}
\usage{
outlier1d(x, return = c("vec", "mat"), minsum = 3)
}
\arguments{
\item{x}{numeric vector}

\item{return}{type to return}

\item{minsum}{1,2 or 3: minimum number of methods
indicating an outlier, only if return is vec}
}
\value{
matrix or vector
}
\description{
Three methods for outlier detection in 1 dimension:
1st and 3rd quartile +/- 1.5*IQR, zscore > 1.5, hampel filter using
median +/- 3*MAD. When return is mat results from all methods are returned.
If return is vec results are collapsed by summation and minsum becomes
relevant.
}
\examples{
x <- c(rnorm(1000), rnorm(20,mean = 10))
outlier1d(x, minsum = 3)
}
