% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_jaccard_index.R
\name{pairwise_jaccard_index}
\alias{pairwise_jaccard_index}
\title{Title}
\usage{
pairwise_jaccard_index(x)
}
\arguments{
\item{x}{named list of vectors}
}
\value{
list of jaccard index formats and plot
}
\description{
Title
}
\examples{
a <- letters[1:5]
b <- letters[1:7]
c <- letters[5:10]
d <- letters[2:8]
x <- list(a=a,b=b,c=c,d=d)
jacdata <- pairwise_jaccard_index(x)
}
