% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot2.R
\name{plot2}
\alias{plot2}
\title{Plot data quickly with scattermore}
\usage{
plot2(
  x,
  color = NULL,
  size = 4,
  palette = NULL,
  colortype = c("d", "c", "discrete", "continuous"),
  legend = c("bottomright", "bottom", "bottomleft", "left", "topleft", "top", "topright",
    "right", "center"),
  transform = NULL,
  least_freq_col_top = FALSE,
  color_text_pos = "tl",
  color_text = NULL,
  discrete_lvls = 8,
  ...
)
}
\arguments{
\item{x}{matrix or data frame, col1 becomes x, col2 becomes y}

\item{color}{name of column for color or .density_r or .density_s for
coloring according to point density raw or scaled between 0 and 1}

\item{size}{dot size}

\item{palette}{color palette}

\item{colortype}{type of color scale}

\item{legend}{where to plot legend, NULL for no legend}

\item{transform}{transform data before plotting, provide function as character,
e.g. "log" or "log10" or "sqrt"}

\item{least_freq_col_top}{plot the color groups in ascending order of
frequency to avoid burying low frequent groups}

\item{color_text_pos}{NULL or c("tl", "tr", "bl", "br"); if NULL not plotted}

\item{color_text}{which string to plot, if NULL color is chosen}

\item{discrete_lvls}{discrete color levels (bins) to split continuous variable
into}

\item{...}{arguments to scattermore::scattermoreplot}
}
\value{
nothing, scattermore plot is plotted
}
\description{
A convenient wrapper around scattermore::scattermoreplot to have easy
color scale and legend.
}
\examples{
brathering::plot2(brathering::points_2d_circ(100))
}
