% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3d.R
\name{plot3d}
\alias{plot3d}
\title{Create a 3D plot}
\usage{
plot3d(
  object,
  x,
  y,
  z,
  color = NULL,
  palette = NULL,
  colortype = c("d", "c", "discrete", "continuous"),
  legend = T,
  size = 2,
  transform = NULL,
  type = c("scatter3d", "mesh3d"),
  backend = c("plotly", "rgl"),
  ...
)
}
\arguments{
\item{object}{matrix or data frame}

\item{x}{name of x column, if missing first column of object used}

\item{y}{name of y column}

\item{z}{name of z column}

\item{color}{name of color column}

\item{palette}{color palette}

\item{colortype}{discrete or continuous color scale}

\item{legend}{plot legend?}

\item{size}{size of points}

\item{transform}{transform data before plotting, provide function as character,
e.g. "log" or "log10" or "sqrt"}

\item{type}{scatter or surface plot, only for plotly backend}

\item{backend}{plotly or rgl plot}

\item{...}{arguments to plotly::plot_ly or rgl::plot3d}
}
\value{
nothing but a plot plotted
}
\description{
Create a 3D plot
}
\examples{
df <- data.frame(
x = rnorm(1000),
y = rnorm(1000),
z = rnorm(1000),
col1 = sample(letters[1:5], 1000, replace = TRUE),
col2 = sample(c(1:10), 1000, replace = TRUE))
plot3d(df, color = ".density_s", colortype = "c", backend = "plotly")
plot3d(df, color = ".density_s", colortype = "c", backend = "rgl")
plot3d(df, color = "col1", colortype = "d", backend = "rgl")
rspheres <- replicate(5, hdos:::random_sphere(), simplify = FALSE)
rtorus <- replicate(5, hdos:::random_torus(), simplify = FALSE)
rcuboid <- replicate(5, hdos:::random_cuboid(), simplify = FALSE)
landscape <- hdos::bind_objects(c(rspheres, rtorus, rcuboid),
                                missing_dim_fill = "runif")
landscape <- hdos::add_noise_points(landscape, density = 0.005)

plot3d(dplyr::bind_cols(landscape), color = "name", legend = FALSE)
plot3d(dplyr::bind_cols(landscape), x = "x2", y = "x3", z = "x4",
       color = "name", legend = FALSE)
plot3d(dplyr::bind_cols(landscape), x = "x3", y = "x4", z = "x5",
       color = "name", legend = FALSE)
plot3d(dplyr::bind_cols(landscape), x = "x4", y = "x5", z = "x6",
       color = "name", legend = FALSE)
}
