% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_filter.R
\name{quantile_filter}
\alias{quantile_filter}
\title{Filter a data frame by quantile limits}
\usage{
quantile_filter(df, valuecol, groupcol = NULL, quantiles = c(0, 1))
}
\arguments{
\item{df}{data frame}

\item{valuecol}{numeric column where to apply quantiles and filter on}

\item{groupcol}{optional grouping column}

\item{quantiles}{numeric of length 2 telling lower and upper quantile limit}
}
\value{
data frame
}
\description{
Filter a data frame by quantile limits
}
\examples{
df <- data.frame(x = rnorm(1000))
# rm lower 10 \% and upper 20 \%
quantile_filter(df = df, valuecol = "x", quantiles = c(0.1, 0.8))
}
