% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_filter.R
\name{quantile_filter2}
\alias{quantile_filter2}
\title{Filter for quantiles across columns}
\usage{
quantile_filter2(x, columns, columns_select = F, quantiles = c(0, 1))
}
\arguments{
\item{x}{data frame or matrix}

\item{columns}{columns in x to quantile filter}

\item{columns_select}{return only columns used for filtering}

\item{quantiles}{lower and upper quantile for filtering}
}
\value{
data frame or matrix
}
\description{
works with matrix and multiple columns simultaneously.
}
