% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_points_2d.R
\name{random_rotation}
\alias{random_rotation}
\title{Random rotation matrix in nD}
\usage{
random_rotation(n)
}
\arguments{
\item{n}{}
}
\description{
Random rotation matrix in nD
}
\examples{
random_rotation(2)
hdos:::is_rotation_matrix(random_rotation(2))
random_rotation(4)
hdos:::is_rotation_matrix(random_rotation(4))
ggplot(ellipse_points_2d(covmat = c(5,0,0,1)), aes(x,y)) +
    geom_point() +
    coord_equal()

ggplot(ellipse_points_2d(covmat = matrix(c(5,0,0,1), nrow = 2) \%*\% random_rotation(2)), aes(x,y)) +
    geom_point() +
    coord_equal()
}
