% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_varname.R
\name{random_varname}
\alias{random_varname}
\title{Generate random variable name that does not appear in x}
\usage{
random_varname(x, prefix = "tmp_", n_digits = 4, max_attempts = 1000)
}
\arguments{
\item{x}{character vector}

\item{prefix}{prefix to return}

\item{n_digits}{random digits}

\item{max_attempts}{loop rounds until n_digits increments by one digit}
}
\value{
character
}
\description{
E.g. for column name of data frame
}
\examples{
df <- data.frame(a=3,b=8,c=12)
df <- df \%>\% dplyr::mutate(!!random_varname(names(.)) := 4)
}
