% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_overlaps.R
\name{range_overlaps}
\alias{range_overlaps}
\title{Find overlapping ranges}
\usage{
range_overlaps(ranges, rm_self = T, rm_mirrored_dup = T)
}
\arguments{
\item{ranges}{(named) list of numeric vectors; can be start end only or complete
range. can even be unordered; only min and max are considered per entry}

\item{rm_self}{remove self overlap from each range}

\item{rm_mirrored_dup}{remove duplicates that only differ in order}
}
\value{
data frame
}
\description{
Find overlapping ranges
}
\examples{
ranges <- list(a = c(1:10),
               b = c(8:20),
               c = c(1:30),
               d = c(40:50))
range_overlaps(ranges)
range_overlaps(ranges, rm_mirrored_dup = F)
range_overlaps(ranges, rm_mirrored_dup = T, rm_self = F)
}
