% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_within.R
\name{reorder_within}
\alias{reorder_within}
\title{Reorder axis text across facets or any column}
\usage{
reorder_within(x, by, within, fun = mean, sep = "___")
}
\arguments{
\item{x}{axis variable}

\item{by}{ordering variable, probably numeric}

\item{within}{group variable, e.g. facet variable}

\item{fun}{what to apply 'by'}

\item{sep}{internal separator}
}
\description{
Reorder axis text across facets or any column
}
\examples{
\dontrun{
ggplot(df, aes(x = reorder_within(x, num_order_col, group))) +
    scale_x_reordered()
}
}
