% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeat_arithmetic.R
\name{repeat_arithmetic}
\alias{repeat_arithmetic}
\title{Repeatedly run arithmetic operation}
\usage{
repeat_arithmetic(x, y = 2, times = 10, fun = `/`)
}
\arguments{
\item{x}{argugment 1}

\item{y}{argument 2}

\item{times}{number of repetitions}

\item{fun}{function}
}
\value{
numeric vector
}
\description{
Have all intermediate results returned.
}
\examples{
repeat_arithmetic(50, 2)
}
