% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_na_all.R
\name{replace_na_all}
\alias{replace_na_all}
\title{Replace all NA values in data frame}
\usage{
replace_na_all(df, value = 0, cols = dplyr::where(is.numeric), exclude = NULL)
}
\arguments{
\item{df}{data frame}

\item{value}{replacement values}

\item{cols}{which columns to scan}

\item{exclude}{columns to exclude}
}
\value{
data frame
}
\description{
Replace all NA values in data frame
}
\examples{
df <- tibble::tibble(
    id       = 1:5,
    score    = c(10, NA, 25, NA, 40),
    name     = c("Alice", NA, "Charlie", "David", NA),
    category = c("A", "B", NA, "A", "C")
)
replace_na_all(df)
replace_na_all(df, value = "0", cols = dplyr::where(is.character))
replace_na_all(df, value = "0", cols = dplyr::where(is.character), exclude = "name")
# error because of mixed column types
# replace_na_all(df, value = "0", cols = dplyr::everything())
}
