% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round2.R
\name{round2}
\alias{round2}
\title{Rounding while preserving sum of elements}
\usage{
round2(x, digits = 0)
}
\arguments{
\item{x}{numeric vector}

\item{digits}{decimal places to round to}
}
\value{
numeric vector
}
\description{
Rounding while preserving sum of elements
}
\examples{
x <- c(21.5539062420449, 19.7605304961694, 20.6575217766391, 20.3202590252772,
20.1906342102765, 21.5574387808329, 19.7095713897403, 19.4116620736685,
20.6269265808014, 21.3010177224109)
sum(x)
sum(round(x,1))
sum(round2(x,1))
}
