% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_default_mofa2.R
\name{run_default_mofa2}
\alias{run_default_mofa2}
\title{Run MOFA2 default}
\usage{
run_default_mofa2(
  list_of_matrices,
  transpose = F,
  outfile = file.path(getwd(), "MOFA2_object.hdf5"),
  interactive_cluster_select = T
)
}
\arguments{
\item{list_of_matrices}{must be feature x sample matrix; so samples (donors) as
columns}

\item{transpose}{if matrices are sample x feature as for PCA, transpose them?}

\item{outfile}{save path for mofa model}

\item{interactive_cluster_select}{get asked about best looking clustering?}
}
\value{
MOFA model object and plots from analysis of single matrices
}
\description{
Run MOFA2 default
}
\examples{
\dontrun{
space <- hdos::random_space(n_spheres = 8, n_tori = 0, n_cuboids = 0, max_dim = 10, n_samples = 20)
# 3 matrices with equal info
s1 = as.matrix(space$coord)
s2 <- s1 + matrix(rnorm(n = ncol(s1)*nrow(s1), sd = 0.5), ncol = ncol(s1))
s3 <- s1 + matrix(rnorm(n = ncol(s1)*nrow(s1), sd = 0.8), ncol = ncol(s1))
liofma <- list(space1 = s1, space2 = s2, space3 = s3)
mo <- run_default_mofa2(liofma, transpose = T)
}
}
