% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2n.R
\name{s2n}
\alias{s2n}
\title{Row wise signal-to-noise}
\usage{
s2n(mat1, mat2, eps = 0.000001)
}
\arguments{
\item{mat1}{numeric matrix1}

\item{mat2}{numeric matrix2}

\item{eps}{epsilon to avoid division by zero}
}
\value{
numeric vector
}
\description{
Row wise signal-to-noise
}
