% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saverds2.R
\name{saverds2}
\alias{saverds2}
\title{Save an R Object as a Gzip-Compressed RDS File}
\usage{
saverds2(x, file, compression = 1)
}
\arguments{
\item{x}{Any R object to serialize.}

\item{file}{A character string specifying the output file path.}

\item{compression}{An integer from `0` to `9` controlling gzip compression.
Higher values generally produce smaller files but take longer. Defaults
to `1`.}
}
\value{
`x`, invisibly.
}
\description{
Serializes an R object using RDS format version 3 and gzip compression,
then reports the output file path.
}
\examples{
path <- tempfile(fileext = ".rds")
saverds2(mtcars, path)
restored <- readr::read_rds(path)
unlink(path)
}
