% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale2.R
\name{scale2}
\alias{scale2}
\title{Scale vector of matrix from min to max value}
\usage{
scale2(x, min = 0, max = 1, margin = 2)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{min}{min value}

\item{max}{max value}

\item{margin}{margin if x is matrix}
}
\value{
numeric vector or matrix
}
\description{
Checkout scales::rescale().
}
\examples{
scale2(rnorm(100), -1, 1)
}
