% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscales.R
\name{scale_color_custom}
\alias{scale_color_custom}
\title{Custom color scale}
\usage{
scale_color_custom(
  colors = colrr::col_pal("custom"),
  name = waiver(),
  na.value = "grey50",
  ...
)
}
\arguments{
\item{colors}{see ggplot2::scale_color_manual}

\item{name}{see ggplot2::scale_color_manual}

\item{na.value}{see ggplot2::scale_color_manual}

\item{...}{see ggplot2::scale_color_manual}
}
\value{
ggplot2::scale_color_manual
}
\description{
Custom color scale
}
