% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscales.R
\name{scale_fill_custom}
\alias{scale_fill_custom}
\title{Custom fill scale}
\usage{
scale_fill_custom(
  colors = colrr::col_pal("custom"),
  name = waiver(),
  na.value = "grey50",
  ...
)
}
\arguments{
\item{colors}{see ggplot2::scale_fill_manual}

\item{name}{see ggplot2::scale_fill_manual}

\item{na.value}{see ggplot2::scale_fill_manual}

\item{...}{see ggplot2::scale_fill_manual}
}
\value{
ggplot2::scale_fill_manual
}
\description{
Custom fill scale
}
