% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reorder_within.R
\name{scale_x_reordered}
\alias{scale_x_reordered}
\title{Fix axis text from reorder_within}
\usage{
scale_x_reordered(..., sep = "___")
}
\arguments{
\item{...}{args to ggplot2::scale_x_discrete}

\item{sep}{internal separator}
}
\description{
Fix axis text from reorder_within
}
