% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfname.R
\name{selfname}
\alias{selfname}
\title{Name a vector by its elements}
\usage{
selfname(x)
}
\arguments{
\item{x}{vector}
}
\value{
named vector
}
\description{
Concise self naming. No magic though.
}
\examples{
selfname(letters)
}
