% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separate2.R
\name{separate2}
\alias{separate2}
\title{Tidyr separate but distinct}
\usage{
separate2(data, col, into, sep, ..., remove = FALSE)
}
\arguments{
\item{data}{data frame}

\item{col}{column to separate}

\item{into}{see tidyr::separate}

\item{sep}{see tidyr::separate}

\item{...}{args to see tidyr::separate}

\item{remove}{see tidyr::separate}
}
\value{
data frame
}
\description{
Separate unique values of the column to be separated. Intended to increase
speed for large data frames with many repeated values in col.
}
\examples{
n <- 1e6
df <- tibble::tibble(
    id   = seq_len(n),
    code = sample(
        sprintf("X\%03d-\%03d", 1:1000, 1:1000),
        size = n,
        replace = TRUE
    )
)

separate_plain <- function(df) {
    df |> tidyr::separate(code, into = c("a", "b"), sep = "-")
}

microbenchmark::microbenchmark(
    plain = separate_plain(df),
    distinct = separate2(df, code, c("a", "b"), "-"),
    times = 5
)
}
