% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq2.R
\name{seq2}
\alias{seq2}
\title{Vectorized version of seq}
\usage{
seq2(from = 1, to = 1, algorithm = c("1", "2"))
}
\arguments{
\item{from}{vector of start values}

\item{to}{vector of end values}

\item{algorithm}{two procedures to choose from, for education only,
use algorithm 1 which is quicker}
}
\value{
list of integer sequences
}
\description{
Vectorized version of seq
}
\examples{
# algorithm 2 is much slower and needs much more memory
fun_alg1 <- seq2(c(1:1000), c(2001:3000), algorithm = 1)
fun_alg2 <- seq2(c(1:1000), c(2001:3000), algorithm = 2)
system.time(replicate(1000, fun_alg1))
system.time(replicate(10, fun_alg2))
}
