% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_cols_rowwise.R
\name{sort_cols_rowwise}
\alias{sort_cols_rowwise}
\title{Sort values across columns}
\usage{
sort_cols_rowwise(df, cols = NULL)
}
\arguments{
\item{df}{data frame}

\item{cols}{columns to sort}
}
\value{
data frame
}
\description{
E.g. to find duplicates which only vary by order.
}
\examples{
df <- data.frame(
    queryHits   = c(5, 22, 3),
    subjectHits = c(2,  1, 8),
    score       = c(0.9, 0.5, 0.7)
)
sort_cols_rowwise(df) # uses first 2 cols by default
}
