% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitAt.R
\name{split_at}
\alias{split_at}
\title{Split a vector at multiple indices}
\usage{
split_at(x, pos)
}
\arguments{
\item{x}{vector to split}

\item{pos}{vector of split indices}
}
\description{
Split a vector at multiple indices
}
\examples{
split_at(x = letters, pos = c(3,10,6))
}
