% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strextract.R
\name{strextract}
\alias{strextract}
\title{Extract regex from string with context}
\usage{
strextract(x, pattern, before = 0, after = 0)
}
\arguments{
\item{x}{character vector of length 1}

\item{pattern}{regular expression}

\item{before}{number of char upstream to extract}

\item{after}{number of char downstream to extract}
}
\value{
character vector
}
\description{
Add context to regex matches in a string.
}
\examples{
txt <- "The quick brown fox jumps over the lazy dog."
strextract(x = txt, pattern = "o[[:alpha:]]{1}", before = 2, after = 2)
}
