% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit2.R
\name{strsplit2}
\alias{strsplit2}
\title{Split character at pattern indices}
\usage{
strsplit2(x, pattern, inds = 1, SIMPLIFY = FALSE)
}
\arguments{
\item{x}{character vector}

\item{pattern}{split pattern in x}

\item{inds}{indices of pattern found in x to use for splitting}

\item{SIMPLIFY}{from mapply, try to return matrix? only works with one inds}
}
\value{
a list
}
\description{
Split character at pattern indices
}
\examples{
x <- c("a_b_c_d_e_f", "q_w_e_r_t")
strsplit2(x, pattern = "_", inds = c(1,2,5), SIMPLIFY = FALSE)
strsplit2(x, pattern = "_", inds = 2, SIMPLIFY = TRUE)
}
