% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit3.R
\name{strsplit3}
\alias{strsplit3}
\title{Split string at pattern maintaining max length of substrings}
\usage{
strsplit3(x, width = 30, sep = "_")
}
\arguments{
\item{x}{character vector}

\item{width}{max width of split elements}

\item{sep}{split symbol to separate at}
}
\value{
list of character vectors
}
\description{
Similar to stringr::str_wrap.
}
\examples{
strsplit3(c("HALLMARK_INTERFERON_ALPHA_RESPONSE", "HALLMARK_INTERFERON_GAMMA_RESPONSE"), 20)
}
