% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_at.R
\name{strsplit_at}
\alias{strsplit_at}
\title{Split string at multiple indices}
\usage{
strsplit_at(x, pos)
}
\arguments{
\item{x}{a character}

\item{pos}{split indices}
}
\value{
vector of characters
}
\description{
Split string at multiple indices
}
\examples{
strsplit_at("Hello my name is", c(2,5,7))
}
