% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg_path_to_coords.R
\name{svg_path_to_coords}
\alias{svg_path_to_coords}
\title{Convert path from Scalable Vector Graphic to cartesian coordinates}
\usage{
svg_path_to_coords(x)
}
\arguments{
\item{x}{svg path}
}
\value{
data frame with xy-coordinates
}
\description{
Convert path from Scalable Vector Graphic to cartesian coordinates
}
\examples{
x <- "M244.5,334.58L246.345,334.678L248.19,334.736L250.035,334.808L251.88,334.7L253.725,334.748L255.57,334.676L257.415,334.682L259.26,334.694L261.105,334.88L262.95,334.892L264.795,335.066L266.64,335.12L268.485,335.222L270.33,335.216L272.175,335.174L274.02,334.94L275.865,334.7L277.71,334.274L279.555,333.968L281.4,333.632L283.245,333.356L285.09,332.822L286.935,332.372L288.78,331.676L290.625,330.782L292.47,329.642L294.315,328.64L296.16,327.452L298.005,326.378L299.85,325.436L301.695,324.584L303.54,323.57L305.385,322.478L307.23,321.392L309.075,320.18L310.92,318.854L312.765,317.414L314.61,316.1L316.455,314.438L318.3,312.956L320.145,311.294L321.99,309.902L323.835,308.294L325.68,306.722L327.525,304.724L329.37,302.9L331.215,300.65L333.06,298.394L334.905,295.94L336.75,293.414L338.595,290.474L340.44,287.408L342.285,284.108L344.13,280.682L345.975,277.028L347.82,273.254L349.665,269.384L351.51,265.352L353.355,261.164L355.2,257.18L357.045,253.226L358.89,249.344L360.735,245.402L362.58,241.856L364.425,238.1L366.27,234.578L368.115,231.098L369.96,227.978L371.805,224.726L373.65,221.666L375.495,218.492L377.34,215.504L379.185,212.36L381.03,209.072L382.875,205.658L384.72,201.998L386.565,197.798L388.41,193.256L390.255,188.114L392.1,182.6L393.945,176.348L395.79,169.652L397.635,162.248L399.48,154.43L401.325,145.832L403.17,136.928L405.015,127.562L406.86,118.112L408.705,108.326L410.55,98.858L412.395,89.702L414.24,81.056L416.085,73.022L417.93,66.218L419.775,60.32L421.62,55.874L423.465,52.952L425.31,51.836L427.155,52.532L429,55.418L430.845,60.068L432.69,66.806L434.535,75.212L436.38,85.298L438.225,96.716L440.07,109.55L441.915,123.248L443.76,137.768L445.605,152.72L447.45,167.738L449.295,182.636L451.14,197.318L452.985,211.442L454.83,224.972L456.675,237.872L458.52,250.028L460.365,261.29L462.21,271.808L464.055,281.318L465.9,290.114L467.745,297.914L469.59,304.856L471.435,311.048L473.28,316.55L475.125,321.116L476.97,325.268L478.815,328.766L480.66,331.646L482.505,334.136L484.35,336.458L486.195,338.246L488.04,339.95L489.885,341.414L491.73,342.71L493.575,343.754L495.42,344.81L497.265,345.572L499.11,346.286L500.955,346.838L502.8,347.312L504.645,347.534L506.49,347.924L508.335,348.194L510.18,348.542L512.025,348.746L513.87,348.998L515.715,348.926L517.56,348.896L519.405,348.77L521.25,348.67"
df <- svg_path_to_coords(x)
}
