% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tagging_system.R
\name{tagging_system}
\alias{tagging_system}
\title{Add tags to files conveniently
Features: Choose or add tags interactively;
Colored tag display (green = assigned);
Option to skip or append to already-tagged videos;
Stores results in a CSV}
\usage{
tagging_system(filevector, output_file = "tags.csv")
}
\arguments{
\item{filevector}{vector of filenames}

\item{output_file}{output csv file}
}
\value{
nothing, file on disk
}
\description{
Add tags to files conveniently
Features: Choose or add tags interactively;
Colored tag display (green = assigned);
Option to skip or append to already-tagged videos;
Stores results in a CSV
}
