% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_removal.R
\name{trail_inds}
\alias{trail_inds}
\title{Get indices of trailing elements of one kind}
\usage{
trail_inds(x, what = 0)
}
\arguments{
\item{x}{vector of whatever}

\item{what}{element to count}
}
\value{
vector of indices
}
\description{
Originally intended to count zeros. So, what = 0 by default.
}
\examples{
trail_inds(rev(c(0,0,1,2,3,4,5)))
trail_inds(rev(c("r","r","r","s","t","u","d","i","o")), "r")
trail_inds(c(1,2,3,4,0,0,0,0))
trail_inds(rev(strsplit("111110001110011001","")[[1]]),"1")
trail_inds(rev(c(NA, NA, NA, 0,0,1,1,2,"a")), NA_character_)
}
