% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer_labels.R
\name{transfer_labels}
\alias{transfer_labels}
\title{Transfer best matching/corresponding labels between two label columns}
\usage{
transfer_labels(df1, df2, labelcol1, labelcol2, joincol = NULL)
}
\arguments{
\item{df1}{data frame 1}

\item{df2}{data frame 2}

\item{labelcol1}{name of column with labels in df1}

\item{labelcol2}{name of column with labels in df2}

\item{joincol}{name(s) of column(s) to join df1 and df2 by; NULL for all
matching colnames in df1 and df2; passed to 'by' of dplyr::inner_join}
}
\value{
list, df1 and df2 with new columns of corresponding labels;
compared is return from compare_labels
}
\description{
Uses compare_labels. Observations in df1 and df2 do not need to match
exactly. Filtering join is applied.
}
\examples{
df1 <- data.frame(
    label1 = c("A","A","A","B","B","C","C","C","C"),
    id = letters[1:9])
df2 <- data.frame(
    label2 = rev(c("X","X","Y","Y","Z","X","Z","Z","Z")),
    id = letters[9:1])
transfer_labels(df1, df2, "label1", "label2", "id")
}
