% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_removal.R
\name{trim_any}
\alias{trim_any}
\title{Trim anything from a vector}
\usage{
trim_any(x, what = 0)
}
\arguments{
\item{x}{vector of whatever}
}
\value{
trimmed vector with attributes
}
\description{
Leading and trailing 'what' is removed from x. Respective counts are
recorded as attributes lead_rm and trail_rm.
}
\examples{
trim_any(c(1,1,1,1,1,1,1), what = 1)
trim_any(c(0,0,0,1,2,3,0,0,0,4,5,0,0))
trim_any(c(NA,NA,NA,1,1,1,NA,NA,2,2,2,"a"), what = NA_character_)
trim_any(as.integer(c(NA,NA,NA,1,1,1,NA,NA,2,2,2)), what = NA_integer_) # not working
}
