% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsne_animation.R
\name{tsne_animation}
\alias{tsne_animation}
\title{Make a tsne movie}
\usage{
tsne_animation(
  x,
  y = NULL,
  max_iters = c(1:9, seq(10, 200, 10), seq(250, 800, 50)),
  fft_tsne_args = list(rand_seed = 42),
  mc.cores = 1,
  out_file = file.path(getwd(), "tsne_anim.mp4"),
  y_col_fun = colrr::scale_color_custom(),
  view_fun = gganimate::view_follow(),
  animate_args = list(nframes = 200, fps = 25, height = 600, width = 800, units = "px",
    device = "png"),
  ffmpeg_renderer_args = list(ffmpeg = "/opt/local/bin/ffmpeg8", format = "mp4", options
    = list(pix_fmt = "yuv420p", preset = "faster", `c:v` = "libx264"))
)
}
\arguments{
\item{x}{numeric matrix}

\item{y}{data frame, the first column of which becomes color variable}

\item{max_iters}{numbers of iteration for tsne}

\item{fft_tsne_args}{args to fft_tsne}

\item{mc.cores}{parallel computing?}

\item{out_file}{output video file}

\item{y_col_fun}{color function for}

\item{animate_args}{args to gganimate::animate}

\item{ffmpeg_renderer_args}{args to gganimate::ffmpeg_renderer}
}
\value{
raw tsne coordinates and gganimate-ggplot for re-rendering
}
\description{
Make a tsne movie
}
\examples{
tsneanim <- tsne_animation(x = hinze@reductions[["pca"]]@cell.embeddings,
                           y = hinze@meta.data[,"celltype_muto",drop = F],
                           mc.cores = 10)
                           # how to zoom smoothly between different time intervals
}
