% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungunzip.R
\name{ungunzip}
\alias{ungunzip}
\title{Unzip gunzip archive on disk}
\usage{
ungunzip(file, out_path = NULL)
}
\arguments{
\item{file}{.gz file}

\item{out_path}{optional output path}
}
\value{
nothing, unpacked files on disk
}
\description{
Simply calls gunzip system command. May only work on unix system.
Unpacked archives appear in same folder by default.
}
\examples{
\dontrun{
ungunzip(file/to/gunzip/file.gz)
}
}
