% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichminmax.R
\name{which.max2}
\alias{which.max2}
\title{Get max indices with ties}
\usage{
which.max2(x, na.rm = F)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{remove NA?}
}
\value{
vector of indices
}
\description{
Instead of first index only as in which.max, all indices
corresponding to max(x) are returned.
}
\examples{
x <- c(1,5,2,5,1)
which.max(x)
which.max2(x)
}
