% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whichminmax.R
\name{which.min2}
\alias{which.min2}
\title{Get min indices with ties}
\usage{
which.min2(x, na.rm = F)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{remove NA?}
}
\value{
vector of indices
}
\description{
Instead of first index only as in which.min, all indices
corresponding to min(x) are returned.
}
\examples{
x <- c(1,5,2,5,1)
which.min(x)
which.min2(x)
}
