% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ceiling_any.R
\name{ceiling2}
\alias{ceiling2}
\title{Ceiling to a given decimal place}
\usage{
ceiling2(x, accuracy = 0.1, f = base::ceiling)
}
\arguments{
\item{x}{numeric vector}

\item{accuracy}{decimal accuracy}

\item{f}{function}
}
\value{
numeric vector
}
\description{
Ceiling to a given decimal place
}
\examples{
ceiling2(rnorm(20))
}
