% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combnn.R
\name{combnn}
\alias{combnn}
\title{Get all element combinations of every lengths, except zero}
\usage{
combnn(
  x,
  order_matters = F,
  repeats_allowed = F,
  min_len = 1,
  consecutive_only = F,
  return_numeric = F
)
}
\arguments{
\item{x}{a vector}

\item{order_matters}{if F: get all combinations; if T: permutations}

\item{repeats_allowed}{allow repeated elements}

\item{min_len}{min combination length of x}

\item{return_numeric}{return values as in x (=F), or as numeric (=T)}
}
\value{
list
}
\description{
The empty element is not included.
Combinations -> subsets, order doesn’t matter, no repetition. This is binomial
coefficients.
Permutations -> subsets, order matters, no repetition. With factorial().
Cartesian product -> subsets, order matters, with repetition. This is expand.grid
of all lengths of x.
}
\examples{
vec <- c("this", "is", "a", "character", "vector")
out <- combnn(vec)
}
