% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_labels.R
\name{compare_labels}
\alias{compare_labels}
\title{Compare two vectors of labels for overlap/correspondence/similarity}
\usage{
compare_labels(
  x,
  y,
  freq_label_cutoff = 0,
  join_label_sep = "_",
  join_label_thresh = 0.05
)
}
\arguments{
\item{x}{vector 1 or data frame; if df then col1 and col2 become x and y}

\item{y}{vector 2}
}
\value{
list
}
\description{
x and y may be two label vectors for same observations. compare how one
splits in the other. In return: raw is pure counts of shared obs; row_props
and col_probs tells relative sharing: e.g row1 of row_props tells the relative
split of label of row1 to collabels, col_probs: same from column perspective.
jaccard is self-explaining. row or colwise max values in row_props or
col_props are row_corres and col_corres. These may be used to transfer
labels.
}
\examples{
# see transfer_labels which uses compare_labels
x <- data.frame(
    label1 = c("A","A","A","B","B","C","C","C","C"),
    label2 = c("X","X","Y","Y","Z","X","Z","Z","Z")
)
compare_labels(x = x)
}
