% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_long_to_mat.R
\name{df_long_to_mat}
\alias{df_long_to_mat}
\title{Convert long data frame to matrix format}
\usage{
df_long_to_mat(df, to_rows = "rname", to_cols = "cname", values = "value")
}
\arguments{
\item{df}{data frame in long format}

\item{to_rows}{column to become rows in matrix}

\item{to_cols}{column to become columns in matrix}

\item{values}{column to populate the matrix}
}
\value{
a matrix
}
\description{
Just for convenience.
}
\examples{
df <- expand.grid(rname = letters[1:5], cname = LETTERS[5:1])
df$value <- 1:25
df_long_to_mat(df)
}
