% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_long_to_mat.R
\name{df_long_to_mat}
\alias{df_long_to_mat}
\title{Convert long data frame to matrix format}
\usage{
df_long_to_mat(df, to_rows = "rows", to_cols = "cols", values = "values")
}
\arguments{
\item{df}{data frame in long format}

\item{to_rows}{column to become rows in matrix}

\item{to_cols}{column to become columns in matrix}

\item{values}{column to populate the matrix}
}
\value{
a matrix
}
\description{
Just for convenience.
}
