% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_points_2d.R
\name{ellipsoid_points_3d}
\alias{ellipsoid_points_3d}
\title{Title}
\usage{
ellipsoid_points_3d(
  mu = c(0, 0, 0),
  covmat = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
  n_theta = 50,
  n_phi = 25,
  c = 1
)
}
\arguments{
\item{c}{}
}
\description{
Title
}
\examples{
brathering::plot3d(ellipsoid_points_3d())
brathering::plot3d(ellipsoid_points_3d(covmat = c(5,0,0, 0,1,0, 0,0,1)))
brathering::plot3d(ellipsoid_points_3d_ran())


brathering::plot3d(ellipsoid_points_3d(covmat = rotate_cov_3d(c(5,0,0, 0,1,0, 0,0,1), Rz = 0)))
# does it work?
brathering::plot3d(ellipsoid_points_3d(covmat = rotate_cov_3d(c(5,0,0, 0,1,0, 0,0,1), Rz = pi)))
}
