% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_points_2d.R
\name{ellipsoid_points_3d_ran}
\alias{ellipsoid_points_3d_ran}
\title{Title}
\usage{
ellipsoid_points_3d_ran(
  mu = c(0, 0, 0),
  covmat = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
  n = 1000,
  c = 1
)
}
\arguments{
\item{c}{}
}
\description{
Title
}
