% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_repel_coords.R
\name{get_repel_coords}
\alias{get_repel_coords}
\title{Get coordinates of repelled labels from ggrepel}
\usage{
get_repel_coords(ggobj, width = 2, height = 2, ...)
}
\arguments{
\item{ggobj}{ggplot object}

\item{width}{passed to grid::viewport, may be subject to iteration to obtain
optimal repelled coordinates}

\item{height}{passed to grid::viewport}

\item{...}{arguments to ggrepel::geom_text_repel}
}
\value{
data frame
}
\description{
Adapted from https://github.com/slowkow/ggrepel/issues/24. Labels are
ploted very late in the plotting process, i.e. their positions are not
pre-computed or so. Hence, it is not straight forward to extract their
repelled positions.
}
\examples{
df <- data.frame(x=rnorm(10), y=rnorm(10), z = letters[1:10])
library(ggplot2)
gg <- ggplot(df, aes(x,y)) +
    geom_point()
gg <-
    gg +
    ggrepel::geom_label_repel(aes(label = z)) +
    geom_text(data = dplyr::left_join(df, get_repel_coords(gg)),
              aes(x = x.repel,y = y.repel, label = z), color = "hotpink")
gg
}
