% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unique_level_columns.R
\name{get_unique_level_columns}
\alias{get_unique_level_columns}
\title{Find columns with unique levels within groups of reference column}
\usage{
get_unique_level_columns(df, refcol)
}
\arguments{
\item{df}{data frame}

\item{refcol}{a column in df, intended for joining to another df}
}
\value{
column names with refcol at first position
}
\description{
It finds columns which can be joined via refcol to another df without
increasing
}
\examples{
df <- tibble::tribble(
    ~id, ~category, ~color,   ~size, ~constantA, ~constantB,
    1,  "A",       "red",    "S",   "foo",      "same",
    2,  "A",       "red",    "M",   "foo",      "same",
    3,  "B",       "blue",   "S",   "bar",      "same",
    4,  "B",       "blue",   "L",   "bar",      "same"
)
joincols <- get_unique_level_columns(df, refcol = "category")
df2 <- tibble::tribble(
    ~category, ~description,         ~priority,
    "A",       "Category A items",   1,
    "B",       "Category B items",   2,
    "C",       "Category C items",   3
)
df3 <- dplyr::left_join(df2,
                        dplyr::distinct(df, dplyr::pick(dplyr::all_of(joincols))),
                        by = "category")
# without selection for joincols, df4 gets longer
df4 <- dplyr::left_join(df2,
                        dplyr::distinct(df),
                        by = "category")
}
