% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_axes_arrows.R
\name{gg_axes_arrows}
\alias{gg_axes_arrows}
\title{Add axes arrows to ggplot}
\usage{
gg_axes_arrows(
  ggobj,
  rel_len = 0.1,
  add_names = T,
  arrow_args = list(angle = 25, length = grid::unit(0.2, "cm"), type = "open"),
  segment_args = list(linewidth = 0.2, color = "black")
)
}
\arguments{
\item{ggobj}{ggplot object}

\item{rel_len}{relative length of axis}

\item{add_names}{add axis titles?}

\item{arrow_args}{arguments to grid::arrow}

\item{segment_args}{arguments to ggplot2::geom_segment}
}
\value{
ggplot object
}
\description{
Add axes arrows to ggplot
}
\examples{
df <- data.frame(x = rnorm(20, 2, 1),
y = rnorm(20, 3, 1))
p <- ggplot(df, aes(x, y)) +
    geom_point()
gg_axes_arrows(p)
}
