% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_save.R
\name{gg_save}
\alias{gg_save}
\title{ggsave with defaults}
\usage{
gg_save(
  plot = ggplot2::last_plot(),
  filename = paste0(deparse(substitute(plot)), ".", deparse(substitute(device))),
  device = png,
  path = getOption("im_path", default = getwd()),
  width = 8,
  height = NULL,
  aspect_ratio = 4/3,
  create.dir = T,
  overwrite = F,
  append_filename = T,
  ...
)
}
\arguments{
\item{plot}{plot}

\item{filename}{filename}

\item{device}{device}

\item{path}{filepath}

\item{width}{w}

\item{height}{h}

\item{aspect_ratio}{set height relative to width: height = width/aspect_ratio}

\item{create.dir}{do it?}

\item{overwrite}{overwrite existing file on disk?}

\item{append_filename}{when overwrite = F, change filename to save plot?}

\item{...}{other arg to ggplot2::ggsave}
}
\value{
nothing
}
\description{
filename on disk derives from plot var name and format (device) by default.
width derive from height and aspect ratio by default - easy scaling.
overwrite management.
path derives from global var by default - reduced redundancy.
}
\details{
set im_path global option: options(im_path = file.path(...))
}
\examples{
tt <- data.frame(x = rnorm(20), y = rnorm(20))
options(im_path = "~/Desktop")

pp <- ggplot2::ggplot(tt, ggplot2::aes(x,y)) +
    ggplot2::geom_point()

gg_save(pp)
}
