% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/img_save.R
\name{img_save}
\alias{img_save}
\title{image save with defaults}
\usage{
img_save(
  plotdata,
  plotfun = plot,
  filename = paste0(deparse(substitute(plotdata)), ".", device),
  device = c("png", "tiff", "jpeg", "bmp"),
  path = getOption("im_path", default = getwd()),
  width = 8,
  height = NULL,
  aspect_ratio = 4/3,
  units = "in",
  dpi = 200,
  overwrite = F,
  append_filename = T,
  device_args = list(),
  plotfun_args = list()
)
}
\arguments{
\item{plotdata}{plotdata}

\item{plotfun}{function for plotting plotdata}

\item{filename}{filename}

\item{device}{device}

\item{path}{filepath}

\item{width}{w}

\item{height}{h}

\item{aspect_ratio}{set height relative to width: height = width/aspect_ratio}

\item{units}{units of height and width}

\item{dpi}{resolution in dots per inch}

\item{overwrite}{overwrite existing file on disk?}

\item{append_filename}{when overwrite = F, change filename to save plot?}

\item{device_args}{args to device}

\item{plotfun_args}{arguments to plotfun}
}
\value{
nothing
}
\description{
filename on disk derives from plot var name and format (device) by default.
width derive from height and aspect ratio by default - easy scaling.
overwrite management.
path derives from global var by default - reduced redundancy.
}
\details{
set im_path global option: options(im_path = file.path(...))
}
\examples{
tt <- data.frame(x = rnorm(20), y = rnorm(20))
options(im_path = "~/Desktop")
img_save(plotdata = tt)
img_save(plotdata = tt, plotfun = brathering::plot2, plotfun_args = list(size = 1))
}
