% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_z_scored.R
\name{is_z_scored}
\alias{is_z_scored}
\title{Test if a vector is z-scored}
\usage{
is_z_scored(x, tol = 0.01)
}
\arguments{
\item{x}{vector}

\item{tol}{tolerance for mean and sd}
}
\value{
logical
}
\description{
Test if a vector is z-scored
}
\examples{
is_z_score(rnorm(100))
}
