% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_index.R
\name{jaccard_index}
\alias{jaccard_index}
\title{Simple jaccard index of elements of 2 vectors}
\usage{
jaccard_index(a, b)
}
\arguments{
\item{a}{vector 1}

\item{b}{vector 2}
}
\value{
a number representing the jaccard index
}
\description{
Simple jaccard index of elements of 2 vectors
}
\examples{
a <- letters[1:5]
b <- letters[1:7]
jaccard_index(a,b)
}
