% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jaccard_index_pairwise.R
\name{jaccard_index_pairwise}
\alias{jaccard_index_pairwise}
\title{Pairwise jaccard indices}
\usage{
jaccard_index_pairwise(x)
}
\arguments{
\item{x}{named list of vectors}
}
\value{
list of jaccard index formats and plot
}
\description{
Pairwise jaccard indices
}
\examples{
a <- letters[1:5]
b <- letters[1:7]
c <- letters[5:10]
d <- letters[2:8]
x <- list(a=a,b=b,c=c,d=d)
jacdata <- jaccard_index_pairwise(x)
}
