% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_chunkwise.R
\name{join_chunkwise}
\alias{join_chunkwise}
\title{Join large number of data frames chunkwise}
\usage{
join_chunkwise(
  df_list,
  join_fun = dplyr::full_join,
  join_by,
  chunks = length(df_list)/10,
  progress = F,
  rm_invalid = F
)
}
\arguments{
\item{df_list}{list of data frames}

\item{join_fun}{function to use for joining; must be one of dplyrs functions}

\item{join_by}{column(s) to join by; passed to "by" arg of join_fun}

\item{chunks}{number of chunks}

\item{progress}{print progress bar?}

\item{rm_invalid}{remove invalid list elements (e.g, no data frame or
no join_by column)}
}
\value{
a single data frame
}
\description{
It is quicker with a reasonable number of chunk.
}
\examples{
# Generate a list of n data frames
install.packages("bench")
ndf <- 200
nrows <- 1000
shared_id <- data.frame(id = 1:nrows)
generate_df <- function(i) {
    n_extra_cols <- sample(2:5, 1)  # Random number of extra columns (2 to 5)
    extra_cols <- replicate(n_extra_cols, rnorm(nrows), simplify = FALSE)
    names(extra_cols) <- paste0("var", i, "_", seq_len(n_extra_cols))
    df <- data.frame(id = shared_id$id, extra_cols)
    return(df)
}
df_list <- lapply(1:ndf, generate_df)

# vary chunk size
benchres <- bench::mark(c1 = join_chunkwise(df_list = df_list, join_by = "id",
                                            chunks = 1),
                        c5 = join_chunkwise(df_list = df_list, join_by = "id",
                                            chunks = length(df_list)/5),
                        c10 = join_chunkwise(df_list = df_list, join_by = "id",
                                             chunks = length(df_list)/10),
                        c20 = join_chunkwise(df_list = df_list, join_by = "id",
                                             chunks = length(df_list)/20),
                        c40 = join_chunkwise(df_list = df_list, join_by = "id",
                                             chunks = length(df_list)/40))
}
