% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_removal.R
\name{lead_count}
\alias{lead_count}
\title{Count number of leading elements of one kind}
\usage{
lead_count(x, what = 0)
}
\arguments{
\item{x}{vector of whatever}

\item{what}{element to count}
}
\value{
integer of count of leading occurrences of what
}
\description{
Use it to count number of leading zero in a time series, e.g. To count
trailing use rev(vec).
}
\examples{
lead_count(c(0,0,1,2,3,4,5))
lead_count(c("r","r","r","s","t","u","d","i","o"), "r")
lead_count(rev(c(1,2,3,4,0,0,0,0)))
lead_count(strsplit("111110001110011001","")[[1]],"1")
lead_count(c(NA, NA, NA, 0,0,1,1,2,"a"), NA_character_)
}
