% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_portable_files.R
\name{make_portable_filepath}
\alias{make_portable_filepath}
\title{Make file paths portable}
\usage{
make_portable_filepath(x, allow = c("-", "."), repl = "_", urldecode = T)
}
\arguments{
\item{x}{vector of file paths}

\item{allow}{values to allow besides alnum}

\item{repl}{replacement for all not permitted values}

\item{urldecode}{also decode URL encoding}
}
\value{
vector of modified paths
}
\description{
This a wrapper around make_portable_filename. Dirname(x) and basename(x)
are treated separately. Dirnames may be subject to url decoding only.
}
