% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_portable_files.R
\name{make_portable_filepath}
\alias{make_portable_filepath}
\title{Make file paths portable}
\usage{
make_portable_filepath(
  x,
  allow = c("-", "."),
  repl = "_",
  urldecode = T,
  make_unique = F,
  repl_empty = "x",
  try_deunicode = F
)
}
\arguments{
\item{x}{character vector of file paths}

\item{allow}{values to allow besides alnum}

\item{repl}{replacement for all not permitted values}

\item{urldecode}{also decode URL encoding}
}
\value{
character vector of modified paths
}
\description{
This a wrapper around make_portable_filename. dirname(x) and basename(x)
are treated separately. dirnames may be subject to url decoding only.
}
\examples{
make_portable_filepath(x = file.path("my", "file", "path",
                                     c("näme1",
                                       "næme2",
                                       "name~3!",
                                       "name!!4",
                                       "...name§5___",
                                       "name/\\\\6",
                                       "name_βγ7",
                                       "name{(8)}",
                                       "name_9\%")))
make_portable_filepath("path\%20withspace/file!!1")
make_portable_filepath("path\%20withspace/file!!1", urldecode = F)
}
