% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mat_to_df_long.R
\name{mat_to_df_long}
\alias{mat_to_df_long}
\title{Matrix to long data frame}
\usage{
mat_to_df_long(
  x,
  rownames_to = "rname",
  colnames_to = "cname",
  values_to = "value"
)
}
\arguments{
\item{values_to}{}
}
\description{
Matrix to long data frame
}
