% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pad_num_in_str.R
\name{pad_num_in_str}
\alias{pad_num_in_str}
\title{Pad one or more numbers in strings to equal length(s)}
\usage{
pad_num_in_str(x, pad = "0", len = "overall_max", len_indwise_remain = 1)
}
\arguments{
\item{x}{character vector}

\item{pad}{pad to stringr::str_pad}

\item{len}{how to pad numbers: overall_max = all to the longest number in x,
indwise_max: index wise longest number, one fixed number to pad all numbers to,
a vector of lengths for each common index of numbers across x}

\item{len_indwise_remain}{how to pad non-common (remaining) numbers in y}
}
\value{
modified x
}
\description{
Intended for subsequent sorting.
}
\examples{
x <- c("w344r5t6", "w3r1t69", "w3r3t7q98")
pad_num_in_str(x)
pad_num_in_str(x, pad = "1")
pad_num_in_str(x, len = 2)
pad_num_in_str(x, len = 5)
pad_num_in_str(x, len = "indwise_max")
pad_num_in_str(x, len = "indwise_max", len_indwise_remain = 5)
}
