% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_pie.R
\name{piechart}
\alias{piechart}
\title{Make pie or donut chart}
\usage{
piechart(
  x,
  order = NULL,
  fill = colrr::col_pal("custom"),
  fill_na = "grey50",
  color = "white",
  color_text = "..auto..",
  radius_inside = 0.3,
  label_outside = c("none", "abs", "rel"),
  label_inside = c("rel", "abs", "none"),
  label_rel_cutoff = 0,
  label_size = 5,
  label_radius_inside = 0.75,
  label_radius_outside = 1.1,
  label_angle_inside = NULL,
  label_angle_outside = NULL,
  label_overlap = c("ignore", "alternate", "outside"),
  overlap_outside_radius = 1.1,
  label_rel_pct = F,
  label_rel_dec = 2,
  legend_title = NULL,
  theme = ggplot2::theme_classic(),
  theme_args = list(panel.grid = ggplot2::element_blank(), axis.title =
    ggplot2::element_blank(), axis.text = ggplot2::element_blank(), axis.ticks =
    ggplot2::element_blank()),
  col_pal_args = list(missing_fct_to_na = T)
)
}
\arguments{
\item{x}{unnamed vector or named numeric vector (= already summarized)}

\item{order}{sort summarized/tabulated groups of x; NULL for no sorting;
TRUE for decreasing, FALSE for increasing}

\item{fill}{color palette to fill pie pieces, can be named by groups in x}

\item{fill_na}{fill for NA}

\item{color}{border color of pie pieces}

\item{color_text}{text color}

\item{radius_inside}{inner radius where pieces are cut. outer radius is fixed
to 1. set to 0 to get pie chart}

\item{label_outside}{which labels outside of pie pieces (radius > 1): none,
absolute, relative}

\item{label_inside}{see label_outside, radius < 1}

\item{label_rel_cutoff}{min. fraction of pie piece to plot label}

\item{label_size}{size of label, passed to geom_text}

\item{label_radius_inside}{vector of radii where to place labels,
gets recycled}

\item{label_radius_outside}{see label_radius_inside}

\item{label_angle_inside}{vector of label angles, gets recycled}

\item{label_angle_outside}{see label_angle_inside}

\item{label_overlap}{method of trying to fix overlapping labels}

\item{overlap_outside_radius}{radius for label_overlap = "outside"}

\item{label_rel_pct}{relative labels in percent?}

\item{label_rel_dec}{decimal places of relative labels}

\item{legend_title}{legend title}

\item{theme}{ggplot theme}

\item{theme_args}{arguments to ggplot2::theme()}

\item{col_pal_args}{args to colrr::make_col_pal}
}
\value{
list of plot and data frame
}
\description{
Make pie or donut chart
}
\examples{
piechart(x = c(rep("a",10), rep("b",6), rep("c", 3), rep("d",2)),
         label_rel_pct = T,
         label_rel_dec = 1)
piechart(x = setNames(1:10, letters[1:10]))
}
