% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recycle.R
\name{recycle}
\alias{recycle}
\title{Recycle vectors or lists}
\usage{
recycle(short, long)
}
\arguments{
\item{short}{shorter element, vector or list}

\item{long}{longer element}
}
\value{
vector or list of length(long) and class(x)
}
\description{
Recycle short element to length of long element.
}
\examples{
# vectors
recycle(letters[1:5], LETTERS[1:15]) # 15/5 = integer
recycle(letters[1:5], LETTERS[1:13]) # 13/5 = double
# vector and list
recycle(list(a = letters[1:5], b = letters[1:3]), LETTERS[1:4])
# lists
recycle(list(a = letters[1:5], b = letters[1:3]), as.list(LETTERS[1:5]))
}
