% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_values_with_names.R
\name{replace_values_with_names}
\alias{replace_values_with_names}
\title{Make function arguments their value}
\usage{
replace_values_with_names(code_block)
}
\arguments{
\item{code_block}{}
}
\value{
character to prep with cat()
}
\description{
Make function arguments their value
}
\examples{
x <- replace_values_with_names(' order = T,
                           fill = colrr::col_pal("custom"),
                           color = "white",
                           radius_inside = 0.3,
                           label_outside = c("none", "abs", "rel"),
                           label_inside = c("rel", "abs", "none"),
                           label_rel_cutoff = 0,
                           label_size = 5,
                           label_radius_inside = 0.75,
                           label_radius_outside = 1.1,
                           label_angle_inside = NULL, # circle or numeric
                           label_angle_outside = NULL, # circle or numeric
                           label_overlap = c("ignore", "alternate", "outside"),
                           overlap_outside_radius = 1.1,
                           label_rel_pct = F,
                           label_rel_dec = 2,
                           legend_title = NULL,
                           theme_args = list(panel.grid = ggplot2::element_blank(),
                                             axis.title = ggplot2::element_blank(),
                                             axis.text = ggplot2::element_blank(),
                                             axis.ticks = ggplot2::element_blank())')
cat(x)
}
