% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_points_2d.R
\name{rotate_cov_2d}
\alias{rotate_cov_2d}
\title{Title}
\usage{
rotate_cov_2d(covmat = c(1, 0, 0, 1), theta = pi * 0)
}
\arguments{
\item{theta}{}
}
\description{
Title
}
\examples{
library(ggplot2)
ggplot(mapping = aes(x,y)) +
    # 90°
    geom_point(data = ellipse_points_2d(mu = c(1,1), covmat = rotate_cov_2d(c(4,0,0,1), pi/2)), color = "firebrick") +
    # 45°
    geom_point(data = ellipse_points_2d(mu = c(1,1), covmat = rotate_cov_2d(c(4,0,0,1), pi/4)), color = "forestgreen") +
    geom_point(data = ellipse_points_2d(mu = c(1,1), covmat = c(4,0,0,1)), color = "cornflowerblue") +
    coord_equal()
}
