% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse_points_2d.R
\name{rotate_cov_3d}
\alias{rotate_cov_3d}
\title{Title}
\usage{
rotate_cov_3d(
  covmat = c(1, 0, 0, 0, 1, 0, 0, 0, 1),
  Rx = 0 * pi,
  Ry = 0 * pi,
  Rz = 0 * pi
)
}
\arguments{
\item{Rz}{}
}
\description{
Title
}
\examples{
rotate_cov_3d()
}
