% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_scale.R
\name{row_scale}
\alias{row_scale}
\title{Title}
\usage{
row_scale(
  x,
  center = TRUE,
  scale = TRUE,
  add_attr = TRUE,
  rows = NULL,
  cols = NULL,
  drop.na.rows = T
)
}
\arguments{
\item{drop.na.rows}{}
}
\description{
Title
}
