% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggscales.R
\name{scale_fill_spectral}
\alias{scale_fill_spectral}
\title{Spectral fill scale}
\usage{
scale_fill_spectral(
  colors = colrr::col_pal("RColorBrewer::Spectral", direction = -1),
  values = NULL,
  name = waiver(),
  na.value = "grey50",
  guide = "colorbar",
  ...
)
}
\arguments{
\item{colors}{see ggplot2::scale_fill_gradient}

\item{values}{see ggplot2::scale_fill_gradient}

\item{name}{see ggplot2::scale_fill_gradient}

\item{na.value}{see ggplot2::scale_fill_gradient}

\item{guide}{see ggplot2::scale_fill_gradient}

\item{...}{see ggplot2::scale_fill_gradient}
}
\value{
ggplot2::scale_fill_gradient
}
\description{
Shortcut function to a spectral fill scale.
}
