% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stack2.R
\name{stack2}
\alias{stack2}
\title{Stack a Data Frame Without Factors}
\usage{
stack2(x, names_to = "name", values_to = "value")
}
\arguments{
\item{x}{A data frame or list of vectors to be stacked.}

\item{names_to}{A character string specifying the name of the output
column that will contain the original column names. Default is `"name"`.}

\item{values_to}{A character string specifying the name of the output
column that will contain the stacked values. Default is `"value"`.}
}
\value{
A data frame with two columns:
\describe{
  \item{<values_to>}{The stacked values from `x`.}
  \item{<names_to>}{The original column names as character (not factor).}
}
}
\description{
A wrapper around [utils::stack()] that returns a long-format data frame
with customizable column names and ensures the variable column is
returned as character (not factor).
}
\details{
This function improves on [utils::stack()] by:
\itemize{
  \item Allowing custom output column names
  \item Returning the indicator column as character instead of factor
}
}
\examples{
df <- data.frame(a = 1:3, b = 4:6)

stack2(df)

stack2(df, names_to = "variable", values_to = "val")
}
