% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_at_num.R
\name{strsplit_at_num}
\alias{strsplit_at_num}
\title{Split string at boundaries of numeric and non-numeric}
\usage{
strsplit_at_num(x)
}
\arguments{
\item{x}{character vector}
}
\value{
list of split characters
}
\description{
First step to extraction or modification of numerical string parts
}
