% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ungunzip.R
\name{ungunzip}
\alias{ungunzip}
\title{Unzip gunzip archive on disk}
\usage{
ungunzip(path)
}
\arguments{
\item{path}{path to .gz file}
}
\value{
nothing, unpacked files on disk
}
\description{
Simply calls gunzip system command. May only work on unix system.
Unpacked archives appear in same folder.
}
\examples{
\dontrun{
ungunzip(path/to/gunzip/file.gz)
}
}
